break up views.py into smaller files
