+++ api/v2/views/__init__.py
@@ -1,0 +1,17 @@
+from .allocation import AllocationViewSet
+from .identity import IdentityViewSet
+from .image import ImageViewSet
+from .image_bookmark import ImageBookmarkViewSet
+from .instance import InstanceViewSet
+from .platform_type import PlatformTypeViewSet
+from .project import ProjectViewSet
+from .project_instance import ProjectInstanceViewSet
+from .project_volume import ProjectVolumeViewSet
+from .provider import ProviderViewSet
+from .provider_machine import ProviderMachineViewSet
+from .provider_type import ProviderTypeViewSet
+from .quota import QuotaViewSet
+from .size import SizeViewSet
+from .tag import TagViewSet
+from .user import UserViewSet
+from .volume import VolumeViewSet
