+++ api/v2/views/image.py
@@ -1,0 +1,16 @@
+from rest_framework import viewsets
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from core.models import Application as Image
+from ..serializers import ImageSerializer
+
+
+class ImageViewSet(viewsets.ModelViewSet):
+    """
+    API endpoint that allows images to be viewed or edited.
+    """
+    queryset = Image.objects.all()
+    serializer_class = ImageSerializer
+    filter_fields = ('created_by__username', 'tags__name')
+    search_fields = ('name', 'description', 'tags__name', 'tags__description')
+    permission_classes = (IsAuthenticatedOrReadOnly,)
+    http_method_names = ['get', 'head', 'options', 'trace']
