+++ api/v2/views/project_instance.py
@@ -1,0 +1,13 @@
+from rest_framework import viewsets
+from core.models import ProjectInstance
+from ..serializers import ProjectInstanceSerializer
+
+
+class ProjectInstanceViewSet(viewsets.ModelViewSet):
+    """
+    API endpoint that allows instance actions to be viewed or edited.
+    """
+    queryset = ProjectInstance.objects.all()
+    serializer_class = ProjectInstanceSerializer
+    filter_fields = ('project__id',)
+    # http_method_names = ['get', 'post', 'delete', 'head', 'options', 'trace']
