+++ api/v2/views/project_volume.py
@@ -1,0 +1,13 @@
+from rest_framework import viewsets
+from core.models import ProjectVolume
+from ..serializers import ProjectVolumeSerializer
+
+
+class ProjectVolumeViewSet(viewsets.ModelViewSet):
+    """
+    API endpoint that allows instance actions to be viewed or edited.
+    """
+    queryset = ProjectVolume.objects.all()
+    serializer_class = ProjectVolumeSerializer
+    filter_fields = ('project__id',)
+    # http_method_names = ['get', 'post', 'delete', 'head', 'options', 'trace']
