+++ api/v2/views/user.py
@@ -1,0 +1,13 @@
+from rest_framework import viewsets
+from core.models import AtmosphereUser
+from ..serializers import UserSerializer
+
+
+class UserViewSet(viewsets.ReadOnlyModelViewSet):
+    """
+    API endpoint that allows users to be viewed or edited.
+    """
+    queryset = AtmosphereUser.objects.all()
+    serializer_class = UserSerializer
+    filter_fields = ('email',)
+    http_method_names = ['get', 'head', 'options', 'trace']
